fix: calculate kernel real size + fix vmm

+++ arch/x86/pmm/pmm.h
@@ -23,7 +23,7 @@ typedef struct {
 } e820_map_t;
 
 void pmm_init(uint64_t kernel_addr, uint64_t kernel_size,
-              uint64_t e820_entries_count, uint32_t e820_entries_addr);
+              uint32_t e820_entries_count, uint32_t e820_entries_addr);
 
 uint32_t pmm_alloc_frame(void);
 uint32_t pmm_alloc_empty_frame(void);
